Added the start_component() function
