+++ docs/api.rst
@@ -9,6 +9,7 @@ Components
 .. autoclass:: Component
 .. autoclass:: ContainerComponent
 .. autoclass:: CLIApplicationComponent
+.. autofunction:: start_component
 
 Concurrency
 -----------
