+++ examples/tutorial1/tests/test_client_server.py
@@ -1,14 +1,15 @@
+# isort: off
 from __future__ import annotations
 
 from collections.abc import AsyncGenerator
 
 import pytest
 from anyio import wait_all_tasks_blocked
+from pytest import CaptureFixture
+from asphalt.core import Context, start_component
+
 from echo.client import ClientComponent
 from echo.server import ServerComponent
-from pytest import CaptureFixture
-
-from asphalt.core import Context
 
 pytestmark = pytest.mark.anyio
 
@@ -17,14 +18,14 @@ pytestmark = pytest.mark.anyio
 async def server(capsys: CaptureFixture[str]) -> AsyncGenerator[None, None]:
     async with Context():
         server = ServerComponent()
-        await server.start()
+        await start_component(server)
         yield
 
 
 async def test_client_and_server(server: None, capsys: CaptureFixture[str]) -> None:
     async with Context():
         client = ClientComponent("Hello!")
-        await client.start()
+        await start_component(client)
         await client.run()
 
     # Grab the captured output of sys.stdout and sys.stderr from the capsys fixture
